% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean_function_d.R
\name{euclidean}
\alias{euclidean}
\title{Find the Greatest Common Divisor (GCD) of two integers.}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{A number.}
}
\value{
The sum of \code{a} and \code{b}.
}
\description{
(b) Description of the algorithm
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)

}
\references{
\href{https://en.wikipedia.org/wiki/Euclidean_algorithm}{Euclidean}
}
