\name{Lab3AdvR-package}
\alias{Lab3AdvR-package}
\alias{Lab3AdvR}
\docType{package}
\title{
\packageTitle{Lab3AdvR}
}
\description{
\packageDescription{Lab3AdvR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Lab3AdvR}
\packageIndices{Lab3AdvR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Lab3AdvR}

Maintainer: \packageMaintainer{Lab3AdvR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
