% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wiki_graph data}
\format{A data frame with 18 rows and 3 variables:
\describe{
  \item{v1}{price, in US dollars}
  \item{v2}{weight of the diamond, in carats}
  \item{w}{ein Text}
}}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\usage{
wiki_graph
}
\description{
A sample dataset using for dijkstra function.
}
\keyword{datasets}
