% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra Algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A data frame.}

\item{init_node}{A numeric scalar.}
}
\value{
Numeric vector, with the shortest distance to each node.
}
\description{
Dijkstra's algorithm is an algorithm for finding the shortest paths between nodes in a graph.
}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
