% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Takes a graph and an initial node and calculates the shortest path from the initial node to every other node in the graph.}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A data frame.}

\item{init_node}{A number.}
}
\value{
The vector of the shortest path to every other node from the starting node.
}
\description{
It is an algorithm used for finding the shortest path from a single node to a single destination node 
by stopping the algorithm onc e the shortest path to the destination node has been determined.
}
\examples{
dijkstra(wiki_graph, 1)
dijkstra(wiki_graph, 3)

}
\references{
\href{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}{Dijkstra}
}
