% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{The Euclidean Algorithm is a technique used for finding the Greatest Common divisor of two integers.}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
The Euclidean Algorithm is a technique used for finding the Greatest Common divisor of two integers.
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)

}
\references{
#' \href{https://en.wikipedia.org/wiki/Euclidean_algorithm}{Euclidean}
}
