% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wiki_graph}
\format{A data frame with 18 rows and 3 variables:
\describe{
  \item{v1}{node}
  \item{v2}{node}
  \item{w}{weight between v1 and v2 nodes}
}}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\usage{
wiki_graph
}
\description{
A dataframe that contains the nodes and the weights between them - a simulation of the graph from wikipedia
}
\keyword{datasets}
